+++ FacebookAudienceNetwork/CHANGELOG.md
@@ -1,9 +1,16 @@
 ## Changelog
-  * 4.27.0.1
-    * Fix an NPE in FacebookInterstitial's listener. 
+  * 4.28.1.1
+    * Enables advanced bidding for all adapters and adds FacebookAdvancedBidder.
+
+  * 4.28.1.0
+    * This version of the adapters has been certified with Facebook Audience Network 4.28.1.
+
+  * 4.28.0.0
+    * This version of the adapters has been certified with Facebook Audience Network 4.28.0.
+	* Removed star rating from the native ad adapter as it has been deprecated by Facebook.
 
   * 4.27.0.0
     * This version of the adapters has been certified with Facebook Audience Network 4.27.0.
 
   * Initial Commit
-  	* Adapters moved from [mopub-android-sdk](https://github.com/mopub/mopub-android-sdk) to [mopub-android-mediation](https://github.com/mopub/mopub-android-mediation/)
+    * Adapters moved from [mopub-android-sdk](https://github.com/mopub/mopub-android-sdk) to [mopub-android-mediation](https://github.com/mopub/mopub-android-mediation/)